backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
pydantic>=2.9.0
pydantic-settings>=2.5.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-groq>=0.1.0
openai>=1.12.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.2
requests>=2.31.0
python-multipart>=0.0.6
reportlab>=4.0.7
aiofiles>=23.2.1
yfinance>=0.2.28
beautifulsoup4>=4.12.0

